refactor(form-fields): split label resolution modules
Separate form field extraction from label resolution internals so future fixes can stay focused on scoring, text heuristics, geometry, or multi-line expansion. The split follows the project cohesion guideline without changing the public buildFormFields API or extraction behavior.
